Refactored task1, added shape package
